+++ packages/edit/client/Place_edit.js
@@ -1,6 +1,11 @@
 
 Kepler.Place.include({
 	
+	getOwner: function() {
+		this._dep.depend();
+		return this.userId ? K.userById(this.userId) : {username: ''}
+	},
+	
 	isEditable: function() {
 		return !!(this.userId ? (K.Profile.id === this.userId) : false);
 	}